Fix adapter collection view setter when collection view is reused across multiple adapters
Summary:
Scenario:

{P56638669}

Fix:
If the collection view being set has been used by a different list adapter since it was last used by this one, treat it as if it is a new collection view.

One optimization we can do is intelligently not reload the map, and only reload the collection view.

Reviewed By: rnystrom

Differential Revision: D3842953

fbshipit-source-id: e22bce41647cd9dc98e360856c1f01db619cb95b