Prevent re-entrant cell querying when dequeuing cells
Summary:
If we hit the image cache in explore, we will end up calling delegate methods that call `-cellForItemAtIndex:itemController:` while in the process of dequeueing a cell. Before fixing that particular issue, I first want to make sure that "re-entrant" (sort of) cell querying is disabled. For starters, I'm just returning `nil` if that happens.

Explore workflow:

1. `-collectionView:cellForItemAtIndexPath:` asks for a cell, routed through the `IGListAdapter`
2. `IGMediaThumbnailItemController` dequeues and then configures a cell
3. When configuring, we set an `NSURL` on the image view
4. The image system makes a cache hit and serially sets the image
5. The view immediately notifies the delegate, which is the **item controller** that the image loaded
6. In the delegate method [[diffusion/FBS/browse/releases%252Frelease-igios-2016.09.08/fbobjc/Apps/Instagram/Instagram/Instagram/Classes/Explore/V3/Controllers/IGMediaThumbnailItemController.m;f44798d5f9d6dd4b5fbbf31eaa4d414d5508d443$174-175|we access the cell]]

I plan on also fixing the error in explore. The cell should be the delegate which has //its own delegate// so it can vend itself. That way there's no cell lookup.

Reviewed By: ryanolsonk

Differential Revision: D3877891

fbshipit-source-id: 2ab9619f6d119f186030482968aecdabbcdb7d92