Add scrollDelegate API for item controllers
Summary: There was a product need for this (cc nicklo) and I'm sure it will be a requested feature come open sourcing. Decided to reorganize the scroll method along with 2 new ones in a new scroll delegate API for item controllers. The only actual IC using this now is stories.

Reviewed By: ocrickard

Differential Revision: D3888426

fbshipit-source-id: 6553759f76a7e9a5fa7e3d038e3d2c5dd54a60a8