Rename to IGListSectionController and codemod
Summary:
- "ItemController" is now "SectionController"
- "Item" (e.g. data models) is now "Object"

Now when using the term "item" in the infra that strictly means a cell (similar to using `indexPath.item`).

This creates a unique name for different parts of the infra where before "item" was used for almost everything. It is also more inline w/ how the infra works w/ UICV.

Reviewed By: ryanolsonk

Differential Revision: D3901830

fbshipit-source-id: 6ae15ae8654dd570e27a98f766f3a696356d4131