Copy object from data source and section controller arrays
Summary:
Saw a random stack trace pop up with an OOB error from enumerating an `__NSArrayM` (which is the private class for a mutable array). It looks like some of the data source methods would end up using and storing the mutable array for section controllers.

Also, if the data source returned a mutable array (most likely they all do) then we would use that array.

Reviewed By: dshahidehpour

Differential Revision: D3943731

fbshipit-source-id: 144df30d34970cb39fde61608d9fe4dca58093e3