Update working range docs
Summary:
Docs were out of sync. Working range used to be an attempt to use a range outside the visible bounds, but that's not the case anymore. Instead it is just a number of section controllers beyond the visible range (much simpler). Updated the docs and graphic to reflect this.

Fixes #84

Reviewed By: dshahidehpour

Differential Revision: D4043045

fbshipit-source-id: bc4ae7a5a822626ffda30ff5ac7a3782a3816027