Add more missing unit tests
Summary:
Working on that coverage %. Hustle to ?

Seriously tho, nearly maxed the coverage that is realistic. I think we need to spend some time on the grid layout tho, there are parts untested that also are unnecessary. I should file a 3.0.0 issue to track work on that.

Waiting for CI ?
Closes https://github.com/Instagram/IGListKit/pull/324

Differential Revision: D4319806

Pulled By: rnystrom

fbshipit-source-id: b626abd65ad73709c231671c0c669414eef1660b