fix #460 by making a copy of the section map
performUpdates was not making any copy of the old map
or the new objects

Due to this, no objects inside the array were getting their
copy() (NSCopying) method called.

This prevented updates to objects in the collectionView.
By making a shallow copy of the map, we trigger the copy()
method of every objects, permitting the comparison of
copied object provided by the user implementation of NSCopying
and the new objects provided by IGListKit Adapter Data Source