Replaced placeholders in changelog
Summary:
Travis is complaining about dead links, so this fixes the one in changelog but there's still a broken link in guides to the core data guide which isn't published yet (via Jazzy).

Should we regen docs but reset everything which isn't the guides? Or are we happy to just leave that one until we release 3.0?
Closes https://github.com/Instagram/IGListKit/pull/649

Differential Revision: D4898644

Pulled By: jessesquires

fbshipit-source-id: d1524ef521931471409021ceb23bea7853628294