Move the DEBUG uniqueness check outside of for-loop inside IGListAdapter
Summary:

The original DEBUG was put inside a for loop, which caused a quadratic looping while creating bigger and bigger NSSet.

This is very inefficient and I verified that by profiling with 10000 objects in the array and it caused significant amount of CPU. The main thread is pretty much unresponsive.

Test Plan:

All test cases pass, verify that no more blocking the main thread by the NSSet creation/dealloc.