Remove XCTAssertThrow in IGListAdapterTests
Summary:
= OSS: =

It's not a good testing strategy to check against the assert, it should be checking the real expected output for a function. It limits our usage that we have to use XCTAssertThrow if there is an assertion throws.

The problem is that `IGAssert` uses the `NSCAssert` which would throw an `NSInternalInconsistentException` which affects all the code path. Another follow up is to investigate how to make IGAssert not crashable.

 ---

Reviewed By: rnystrom

Differential Revision: D7481905

fbshipit-source-id: 24af37a6e1aa389317474614cdd214c9df537291