Fix bug with cell identifiers
Summary:
When we added the ability to register the same class with multiple reuse identifiers, forgot to adjust the `registeredClasses` set to reflect that. This resulted in a bug where the collection view could be asked for a cell of a class that had already been registered, but with a different reuse identifier, which would end in a crash.

This diff fixes the crash, I'll follow up with another diff that adds unit tests around this, but want to fix master ASAP.

Differential Revision: D9560338

fbshipit-source-id: e8df6fe5d9929b44842c6abc01d52064ccb4147e