Add test for IGListAdapter dequeue method
Summary: Adding a test to catch the bug that caused the crash in T33376364. In that bug, dequeueing two cells of the same class, but with different reuse identifiers, would cause a crash. Here in the test, if the same bug were to occur the dequeuing of the second cell would throw an exception and the test would fail.

Differential Revision: D9560911

fbshipit-source-id: 9310839aaf441e919b5d443feaa44c71e991d146