Gate the preferItemReloadsForSectionReloads only to use when there is no change to the number of sections
Summary:
There are some crashes related to this path when the number of sections are changed when using the `preferItemReloadsForSectionReloads` optimization for item reloads.

For now, this change would just limit the flag only useable when the section count is unchanged. The proper fix would be finding a repro-case for the unsafe trace.
The follow-up here is adding more logging to capture the crash reason and investigate any variable that leads to this crash.

Reviewed By: calimarkus

Differential Revision: D12828061

fbshipit-source-id: 040e66c051c54f889ee9fbc3e4a48ab6bf93e162