Add experiment to perform IGListAdapterUpdater background updates on a serial queue.
Summary: This diff adds a new `IGListExperimentBackgroundDiffingSerial` experiment that behaves similarly to `IGListExperimentBackgroundDiffing` but executes changes in a per-`IGListAdapterUpdater` serial queue instead of a global queue. Global queues are concurrent, and concurrent queues by nature don't ensure the execution order of the blocks that are enqueued into them, so there's a chance that the former `IGListExperimentBackgroundDiffing` experiment would lead to changes being applied out of order.

Reviewed By: calimarkus

Differential Revision: D13567403

fbshipit-source-id: a2aebb20e5e7dc22601fad3c2dfc80112747c81b