Remove the unnecessary assert which would affect IGListKit perf
Summary:
This assert was hoping to find if there are any duplicates from the `fromObjects`. However, from the finding, we did not see any assertion or trace that this is the case.

The `objectsWithDuplicateIdentifiersRemoved` is actually doing extra work during updates.

We should remove this for the sake of better scroll perf and faster UI update.

Reviewed By: calimarkus

Differential Revision: D14339124

fbshipit-source-id: ff40fd23a05640058673d46f63a9e4bf516dd3f6