Use script to fetch version for podspec
Summary: This allows us to have a single declaration of the version in the `Info.plist`, and have it automatically synced to the podspec. I'm going to split the podspecs as I did it wrong (#1382) when splitting the libraries, so this avoids having three places where the version is declared.

Differential Revision: D18449842

fbshipit-source-id: f49891f6eeb36c2cae70b5338cea96b2e121ef31